interface: Add `get_config_data` helper
#### Problem

The auto-generated clients aren't ready for SDK v3, and Agave is
currently using some auto-generated clients functions for the CLI. The
client libraries currently create a circular dependency with Agave which
we need to break.

#### Summary of changes

Agave can be transitioned to `solana-config-interface` for everything
except the `get_config_data` helper, so add that missing helper.