CI: Move repo to a (newer) Makefile (#63)
* CI: Move repo to a (newer) Makefile

#### Problem

Although the config repo is already using a Makefile, it is still
using its own github workflows and scripts.

#### Summary of changes

Unify everything by using the Makefile and scripts from the ALT repo.
Similar to others, this removes the `bpf-entrypoint` and `test-sbf`
features, renames the program-mb.so, use the interface crate in benches,
add words to the dictionary as needed.

* Update ring, restore lib target for program

* Add dependabot

* Get program names right for publish job

* Add --all-features to clippy