program: Update to sdk v3 crates
#### Problem

The program is still on sdk v2, but sdk v3 is out.

#### Summary of changes

Update dependencies, and use split-up dependencies while we're at it.
The new version reduces CUs for a good amount of instructions!

The client crate needs to stay on v2 because we eliminated borsh v0.10
support in v3, and kaigan requires v0.10 of borsh until
https://github.com/metaplex-foundation/kaigan/pull/9 lands.